# [Merged by Bors] - activation: remove read lock from GetAtxHeader