## Motivation
The exclusive lock is acquired in StoreAtx, which may lead to contention if GetAtxHeader is called concurrently many times.
The difference in benchmark is significant:

```
benchmark                                          old ns/op     new ns/op     delta
BenchmarkGetAtxHeaderWithConcurrentStoreAtx-16     68058         13280         -80.49%
```

related: https://github.com/spacemeshos/go-spacemesh/issues/2478

## Changes
Read lock is not acquired when reading a single key from the database. 


## Test Plan
unit tests

## TODO
<!-- This section should be removed when all items are complete -->
- [x] Explain motivation or link existing issue(s)
- [x] Test changes and document test plan
- [x] Update documentation as needed

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)